Refactor TEC-1G LCD controller
